+++ egui/src/widgets/plot/items.rs
@@ -915,7 +915,7 @@ impl PlotItem for Text {
             .layout_no_wrap(self.text.clone(), self.style, color);
         let rect = self
             .anchor
-            .anchor_rect(Rect::from_min_size(pos, galley.size));
+            .anchor_rect(Rect::from_min_size(pos, galley.size()));
         shapes.push(Shape::galley(rect.min, galley));
         if self.highlight {
             shapes.push(Shape::rect_stroke(