+++ egui_demo_lib/src/apps/http_app.rs
@@ -361,7 +361,7 @@ impl ColoredText {
         let mut job = self.0.clone();
         job.wrap_width = ui.available_width();
         let galley = ui.fonts().layout_job(job);
-        let (response, painter) = ui.allocate_painter(galley.size, egui::Sense::hover());
+        let (response, painter) = ui.allocate_painter(galley.size(), egui::Sense::hover());
         painter.add(egui::Shape::galley(response.rect.min, galley));
     }
 }