+++ epaint/src/shape.rs
@@ -161,7 +161,7 @@ impl Shape {
         color: Color32,
     ) -> Self {
         let galley = fonts.layout_no_wrap(text.to_string(), text_style, color);
-        let rect = anchor.anchor_rect(Rect::from_min_size(pos, galley.size));
+        let rect = anchor.anchor_rect(Rect::from_min_size(pos, galley.size()));
         Self::galley(rect.min, galley)
     }