+++ epaint/src/tessellator.rs
@@ -621,7 +621,7 @@ impl Tessellator {
                 if options.debug_paint_text_rects {
                     self.tessellate_rect(
                         &PaintRect {
-                            rect: Rect::from_min_size(text_shape.pos, text_shape.galley.size)
+                            rect: Rect::from_min_size(text_shape.pos, text_shape.galley.size())
                                 .expand(0.5),
                             corner_radius: 2.0,
                             fill: Default::default(),